// package/admin/src/graphql/schema.types.ts
export type Scalars = {
    ID: { input: string; output: string };
    String: { input: string; output: string };
    Boolean: { input: boolean; output: boolean };
    Int: { input: number; output: number };
    Float: { input: number; output: number };
    DateTime: { input: any; output: any };
};

export type BlogPost = {
    category: Category;
    categoryId: Scalars["ID"]["output"];
    content: Scalars["String"]["output"];
    createdAt: Scalars["DateTime"]["output"];
    id: Scalars["ID"]["output"];
    status: PostStatus;
    title: Scalars["String"]["output"];
    updatedAt: Scalars["DateTime"]["output"];
};
// ...
export type Category = {
    blogPosts: CategoryBlogPostsConnection;
    createdAt: Scalars["DateTime"]["output"];
    id: Scalars["ID"]["output"];
    title: Scalars["String"]["output"];
    updatedAt: Scalars["DateTime"]["output"];
};

export type CategoryBlogPostsArgs = {
    filter?: BlogPostFilter;
    paging?: OffsetPaging;
    sorting?: Array<BlogPostSort>;
};

export type CategoryBlogPostsConnection = {
    /** Array of nodes. */
    nodes: Array<BlogPost>;
    /** Paging information */
    pageInfo: OffsetPageInfo;
    /** Fetch total count of records */
    totalCount: Scalars["Int"]["output"];
};
// ...
export type OffsetPageInfo = {
    /** true if paging forward and there are more records. */
    hasNextPage?: Maybe<Scalars["Boolean"]["output"]>;
    /** true if paging backwards and there are more records. */
    hasPreviousPage?: Maybe<Scalars["Boolean"]["output"]>;
};